requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.1.1
googlemaps==4.6.0
gunicorn==20.1.0
haversine==2.7.0
idna==3.4
numpy==1.23.4
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
requests==2.28.1
reverse-geocode==1.4.1
scipy==1.9.3
six==1.16.0
sqlparse==0.4.3
toposort==1.7
urllib3==1.26.12
whitenoise==6.2.0
